Rename enhancers to conditions
